+++ fiddles/bash/fiddle-0081-Docker/bin/brew/_install_docker.sh
@@ -1,0 +1,11 @@
+#!/usr/bin/env bash
+
+function brewInstallDocker() {
+    groupLog "brewInstallDocker";
+    brew install docker || exit $?;
+}
+
+function brewInstallDockerMachine() {
+    groupLog "brewInstallDockerMachine";
+    brew install docker-machine || exit $?;
+}