+++ fiddles/bash/fiddle-0081-Docker/bin/brew/_install_virtualbox.sh
@@ -1,0 +1,6 @@
+#!/usr/bin/env bash
+
+function brewInstallVirtualbox() {
+    groupLog "brewInstallVirtualbox";
+    brew cask install virtualbox || exit $?;
+}