+++ CHANGELOG.markdown
@@ -965,11 +965,12 @@ Alternately, the _Daily Diary of my Programming Dreams_.
            9.  Updated [scripts/fiddle-start.sh](scripts/fiddle-start.sh)
 
 * **Aurelia**
-
+    1.  Added [fiddles/aurelia/fiddle-0000-Template](fiddles/aurelia/fiddle-0000-Template)
+    
 * **Bash**
     1.  Added [fiddles/bash/fiddle-0079-Aurelia](fiddles/bash/fiddle-0079-Aurelia)
     2.  Added [fiddles/bash/fiddle-0080-Ember](fiddles/bash/fiddle-0080-Ember)
+    3.  Added [fiddles/bash/fiddle-0081-VirtualBox](fiddles/bash/fiddle-0081-VirtualBox)
 
 * **Ember**
     1.  Added [fiddles/ember/fiddle-0001-SuperRental](fiddles/ember/fiddle-0001-SuperRental)
-* Added [fiddles/aurelia/fiddle-0000-Template](fiddles/aurelia/fiddle-0000-Template)